Default secret.py file for WiFi SSID and password.
This file contains the default WiFi SSID and password for applications
in this directory.  Note you absolutely should change the values and
never check it back in.  If you are creating a public access point, or
an access point for PICO use only, you may wish to change these values
to something different.