Adding a simple tictactoe program.
Tic Tac Toe designed for the PICO with a PICO_DISPLAY on it.  This
simple program is meant to be an example as to how to make a small game
work.  There is also the possibility to expand the code so that the
computer move function is more intellegent.  One way to do this would be
to utilize the list of winning moves and determine if there is somehting
that is already and *almost* winning move.