Modify code to have an "AI" powered move selector
This code provides an "AI" like move selector, based on pattern
recongnition of boards.  Overall the objective is that the computer
never loses... it can tie, but it can't lose.

The LED glows red when the computer selects a random move, and blue for
when it finds a pattern to use.  Green indicates the game is over.