Updated to be a shell command.
Modified the script to that we include the /usr/bin/python3 at the
begining so users do not have to remember to run the command with
"python3" first.  Also updated the help messages and error processing a
little so that if theere is no firmware file mentioned, the program
shuts down nicely.